fix 'entity' package naming issue
