+++ src/main/java/com/seailz/javadiscordwrapper/DiscordJv.java
@@ -4,21 +4,11 @@ import com.seailz.javadiscordwrapper.events.DiscordListener;
 import com.seailz.javadiscordwrapper.events.EventDispatcher;
 import com.seailz.javadiscordwrapper.events.annotation.EventMethod;
 import com.seailz.javadiscordwrapper.events.model.interaction.button.ButtonInteractionEvent;
-import com.seailz.javadiscordwrapper.events.model.interaction.select.StringSelectMenuInteractionEvent;
-import com.seailz.javadiscordwrapper.events.model.interaction.select.entity.ChannelSelectMenuInteractionEvent;
-import com.seailz.javadiscordwrapper.events.model.interaction.select.entity.RoleSelectMenuInteractionEvent;
-import com.seailz.javadiscordwrapper.events.model.interaction.select.entity.UserSelectMenuInteractionEvent;
 import com.seailz.javadiscordwrapper.gateway.GatewayFactory;
 import com.seailz.javadiscordwrapper.model.application.Application;
 import com.seailz.javadiscordwrapper.model.channel.Channel;
-import com.seailz.javadiscordwrapper.model.channel.utils.ChannelType;
 import com.seailz.javadiscordwrapper.model.component.ActionRow;
 import com.seailz.javadiscordwrapper.model.component.button.Button;
-import com.seailz.javadiscordwrapper.model.component.select.SelectOption;
-import com.seailz.javadiscordwrapper.model.component.select.entitiy.ChannelSelectMenu;
-import com.seailz.javadiscordwrapper.model.component.select.entitiy.RoleSelectMenu;
-import com.seailz.javadiscordwrapper.model.component.select.entitiy.UserSelectMenu;
-import com.seailz.javadiscordwrapper.model.component.select.string.StringSelectMenu;
 import com.seailz.javadiscordwrapper.model.guild.Guild;
 import com.seailz.javadiscordwrapper.model.application.Intent;
 import com.seailz.javadiscordwrapper.model.user.User;
