+++ src/main/java/com/seailz/javadiscordwrapper/events/model/interaction/select/entity/UserSelectMenuInteractionEvent.java
@@ -2,8 +2,6 @@ package com.seailz.javadiscordwrapper.events.model.interaction.select.entity;
 
 import com.seailz.javadiscordwrapper.DiscordJv;
 import com.seailz.javadiscordwrapper.events.model.interaction.InteractionEvent;
-import com.seailz.javadiscordwrapper.model.component.select.SelectOption;
-import com.seailz.javadiscordwrapper.model.component.select.string.StringSelectMenu;
 import com.seailz.javadiscordwrapper.model.interaction.data.message.MessageComponentInteractionData;
 import com.seailz.javadiscordwrapper.model.user.User;
 import org.jetbrains.annotations.NotNull;
@@ -30,7 +28,7 @@ public class UserSelectMenuInteractionEvent extends InteractionEvent {
     }
 
     /**
-     * Returns the selected users of the {@link com.seailz.javadiscordwrapper.model.component.select.entitiy.UserSelectMenu UserSelectMenu}.
+     * Returns the selected users of the {@link com.seailz.javadiscordwrapper.model.component.select.entity.UserSelectMenu UserSelectMenu}.
      *
      * @throws IllegalStateException if the event was not fied in a {@link com.seailz.javadiscordwrapper.model.guild.Guild Guild}.
      *
